requirements.txt
# pytorch
torch

# langgraph
langgraph

# langchain
langchain
langchain-core
langchain-community


# sqlite
langgraph-checkpoint-sqlite


# OpenAI
langchain-openai
openai

# huggingface
langchain-huggingface
transformers
huggingface-hub

# ḍotenv
python-dotenv

# utils
numpy
pandas
typing
pydantic
requests
ipython
streamlit